feat: Add get hangout functionality to controller
