+++ YoLetsChillBackend/src/main/kotlin/com/dave/yo_lets_chill_backend/controller/contract/GetHangoutRequestyBody.kt
@@ -1,0 +1,5 @@
+package com.dave.yoletschillbackend.controller.contract
+
+data class GetHangoutRequestBody (
+    val password: String?
+)
